requirements.txt
alpaca-py == 0.10.0
alpaca_trade_api == 3.0.2
pandas == 2.1.0
yfinance == 0.2.28
Django == 4.2.5

